Initial public release of cyclops
Cyclops is a software library for monocular visual-inertial odometry
(MVIO) of a camera/IMU system. This software library is supplementary
material for an article we submitted for review to IEEE Transactions on
Robotics.

Cyclops features very fast and accurate initialization compared to other
MVIO algorithms, e.g. ORB-SLAM 3 and VINS-Mono. This fast initialization
is based on almost-exact 1D approximation of MVIO optimization, as
detailed in our submitted paper.